# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yunora AI Question Generation Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects
      security:
        - bearerAuth: []
      parameters:
        - name: standardId
          in: query
          schema: { type: integer }
        - name: boardId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectList"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update subject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete subject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── CHAPTERS ─────────────────────────────────────────────────────────────────
...
  /topics:
    get:
      operationId: listTopics
      tags: [topics]
      summary: List topics
      security:
        - bearerAuth: []
      parameters:
        - name: chapterId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of topics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicList"
    post:
      operationId: createTopic
      tags: [topics]
      summary: Create a topic
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
  /topics/{id}:
    patch:
      operationId: updateTopic
      tags: [topics]
      summary: Update topic
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
    delete:
      operationId: deleteTopic
      tags: [topics]
      summary: Delete topic
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── QUESTION TYPES ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Subject:
      type: object
      required: [id, name, standardId]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: ["string", "null"] }
        standardId: { type: integer }
        standardName: { type: ["string", "null"] }
        boardId: { type: integer }
        boardName: { type: ["string", "null"] }
        isActive: { type: boolean }
        chaptersCount: { type: integer }
        questionsCount: { type: integer }
        createdAt: { type: string, format: date-time }
    SubjectInput:
      type: object
      required: [name, standardId]
      properties:
        name: { type: string }
        code: { type: string }
        standardId: { type: integer }
        isActive: { type: boolean, default: true }
    SubjectUpdate:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
        isActive: { type: boolean }
    SubjectList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subject"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── CHAPTERS ──────────────────────────────────────────────────────────────
...
    Topic:
      type: object
      required: [id, name, chapterId]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        chapterId: { type: integer }
        chapterName: { type: ["string", "null"] }
        isActive: { type: boolean }
        questionsCount: { type: integer }
        createdAt: { type: string, format: date-time }
    TopicInput:
      type: object
      required: [name, chapterId]
      properties:
        name: { type: string }
        description: { type: string }
        chapterId: { type: integer }
        isActive: { type: boolean, default: true }
    TopicUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        isActive: { type: boolean }
    TopicList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Topic"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── QUESTION TYPES ────────────────────────────────────────────────────────
...